pi_kvm_bridge/requirements.txt
Flask
evdev
requests
pyserial
